restructured torch and tensorflow extractor + fixed memory leak/slowdown
